refactor: update ParticipantForm to apply new labels
